fix: Fix exclusive transformers
