+++ Makefile
@@ -30,6 +30,7 @@ PKG        := ./...
 LDFLAGS    := -w -s
 
 SRC        = $(shell find . -type f -name '*.go' -print)
+ASSETS     = $(shell find assets -type f -name '*' -print)
 ARCH       = $(shell uname -p)
 GIT_COMMIT = $(shell git rev-parse HEAD)
 GIT_SHA    = $(shell git rev-parse --short HEAD)
@@ -71,7 +72,7 @@ help: ## This help.
 build: get $(BINDIR)/$(BINNAME) ## Build go code
 	@printf "\033[32m-------------------------------------\n BUILD SUCCESS\n-------------------------------------\033[0m\n"
 
-$(BINDIR)/$(BINNAME): $(SRC)
+$(BINDIR)/$(BINNAME): $(SRC) $(ASSETS)
 	go build -ldflags '$(LDFLAGS)' -o $(BINDIR)/$(BINNAME) ./cmd/${BINNAME}
 ifeq ($(HAS_UPX),true)
 	@echo 'upx detected. compressing binary...'
