+++ assets/filepermissions.yaml
@@ -85,3 +85,6 @@
 "inbuilt/transformers/generators/knative/knative.yaml" : 0644
 "inbuilt/transformers/generators/kubernetes/kubernetes.yaml" : 0644
 "inbuilt/transformers/generators/tekton/tekton.yaml" : 0644
+"inbuilt/transformers/scriptsgenerator/containerpushscripts/containerpushscripts.yaml" : 0644
+"inbuilt/transformers/scriptsgenerator/containerpushscripts/templates/pushimages.bat" : 0755
+"inbuilt/transformers/scriptsgenerator/containerpushscripts/templates/pushimages.sh" : 0755
