+++ assets/inbuilt/transformers/cnb/cfcnb/cfcnb.yaml
@@ -7,8 +7,6 @@ spec:
   class: "CNBContainerizer"
   generates: 
     - "ContainerBuild"
-    - "IR"
-    - "CNBMetadata"
   exclusive: 
     - "ContainerBuild"
   config:
