+++ assets/inbuilt/transformers/cnb/cnbgenerator/cnbgenerator.yaml
@@ -5,7 +5,5 @@ metadata:
 spec:
   mode: "Container"
   class: "CNBGenerator"
-  generates: 
-    - "ContainerBuildScript"
   consumes: 
     - "CNBMetadata"
