+++ assets/inbuilt/transformers/executable/dockerfile/golang/golang.yaml
@@ -7,10 +7,8 @@ spec:
   class: "SimpleExecutable"
   generates: 
     - "ContainerBuild"
-    - "Dockerfile"
   exclusive: 
     - "ContainerBuild"
-    - "Dockerfile"
   config:
     directoryDetectCMD: ["./m2kdetect.sh"]
     container:
