+++ assets/inbuilt/transformers/executable/dockerfile/java-war-tomcat/java-war-tomcat.yaml
@@ -7,9 +7,7 @@ spec:
   class: "SimpleExecutable"
   generates: 
     - "ContainerBuild"
-    - "Dockerfile"
   exclusive: 
     - "ContainerBuild"
-    - "Dockerfile"
   config:
     directoryDetectCMD: ["./m2kdetect.sh"]
