+++ assets/inbuilt/transformers/executable/dockerfile/javagradle/javagradle.yaml
@@ -7,9 +7,7 @@ spec:
   class: "SimpleExecutable"
   generates: 
     - "ContainerBuild"
-    - "Dockerfile"
   exclusive: 
     - "ContainerBuild"
-    - "Dockerfile"
   config:
     directoryDetectCMD: ["./m2kdetect.sh"]
