+++ assets/inbuilt/transformers/executable/dockerfile/python/python.yaml
@@ -7,9 +7,7 @@ spec:
   class: "SimpleExecutable"
   generates: 
     - "ContainerBuild"
-    - "Dockerfile"
   exclusive: 
     - "ContainerBuild"
-    - "Dockerfile"
   config:
     directoryDetectCMD: ["./m2kdetect.sh"]
