+++ assets/inbuilt/transformers/executable/s2i/python/python.yaml
@@ -7,9 +7,7 @@ spec:
   class: "SimpleExecutable"
   generates: 
     - "ContainerBuild"
-    - "S2IBuilder"
   exclusive: 
     - "ContainerBuild"
-    - "S2IBuilder"
   config:
     directoryDetectCMD: ["./m2kdetect.sh"]
