+++ assets/inbuilt/transformers/scriptsgenerator/containerpushscripts/containerpushscripts.yaml
@@ -1,0 +1,11 @@
+apiVersion: move2kube.konveyor.io/v1alpha1
+kind: Transformer
+metadata:
+  name: ContainerImagePushScript
+spec:
+  mode: "Container"
+  class: "ContainerImagePushScript"
+  generates: 
+    - "ContainerImagePushScript"
+  consumes: 
+    - "NewImage"
